fix: move withLogger to IO.ts
